requirements.txt
transformers
tqdm 
python-dotenv
datasets
torch # Used if running models locally
chromadb